xml: another cache_associativity fix

+++ hwloc/topology-xml.c
@@ -155,7 +155,7 @@ hwloc__xml_import_object_attr(struct hwloc_topology *topology, struct hwloc_obj 
   }
 
   else if (!strcmp(name, "cache_associativity")) {
-    unsigned long lvalue = atoi(value);
+    int lvalue = atoi(value);
     if (hwloc_obj_type_is_cache(obj->type))
       obj->attr->cache.associativity = lvalue;
     else if (hwloc__xml_verbose())
